backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
httpx==0.25.1
aiosqlite==0.19.0
pydantic==2.5.2
python-dotenv==1.0.0
py-clob-client==0.34.6
eth-account>=0.10.0
